Move team page to Root, replace raw <a> with <Link>
