+++ src/app/common/components/AppRoot.tsx
@@ -966,7 +966,7 @@ export default class extends Root<Props, State, SharedState, Events> {
 					onViewPrivacyPolicy: this._viewPrivacyPolicy,
 					stripe: this.props.stripeLoader.value
 				}
-			)
+			),
 		};
 
 		// state
