+++ src/app/common/components/TeamPage.scss
@@ -1,9 +1,9 @@
 /*
-        componentPath: src/app/common/components/BrowserRoot/TeamPage
-        cssComponentId: team-page_a0fwiy
+        componentPath: src/app/common/components/TeamPage
+        cssComponentId: team-page_a9b9l1
 */
 
-.team-page_a0fwiy {
+.team-page_a9b9l1 {
 	.home-panel_w65z7c:nth-child(odd) {
 		background-color: var(--app-background-alt-color);
 	}
