+++ src/app/common/components/TeamPage/Profile.scss
@@ -3,7 +3,7 @@
         cssComponentId: profile_m7zor
 */
 
-@import '../../../../../../src/common/styles/breakpoints.scss';
+@import '../../../../../src/common/styles/breakpoints.scss';
 
 
 $tablet-up-margin-title-intro: 0.8rem;
